% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{calculateOutlierScoreByCluster}
\alias{calculateOutlierScoreByCluster}
\title{Calculate cell outlier scores using isolation forests by group}
\usage{
calculateOutlierScoreByCluster(
  sce,
  clusterVar,
  dimred = NULL,
  use_pcs = FALSE,
  plot = TRUE,
  prediction_thresh = 0.5,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object containing a logcounts matrix}

\item{clusterVar}{the name of the cluster variable in the colData of the sce
input.}

\item{dimred}{if specified, plot the specified reduced dimension with
\link[scater]{plotReducedDim}, colored by outlier score}

\item{use_pcs}{if TRUE, run the outlier detection in principal component
space. Otherwise, use logcounts.}

\item{plot}{if TRUE and a dimred is specified, plot the isoscore against the
specified dimred}

\item{prediction_thresh}{threshold to use for binary outlier calling}

\item{...}{arguments to pass to \link[isotree]{isolation.forest}}
}
\description{
This function calculates outlier scores for each cell BY CLUSTER
  using \link[isotree]{isolation.forest}. See [calculateOutlierScore()] for
  more detail on isoforests. This function calls that one on each cluster
  separately. So instead of outlierness from the full dataset overall, this
  function captures outlierness from the cell's cluster.
}
\details{
This function treats NAs in clusterVar as a separate level and
  applies the outlier scoring to them accordingly. The isolation forest model
  objects are returned as a list in the metadata.
}
\examples{
# TODO update this example.
library(scater)
library(scran)
library(scRNAseq)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
query_data <- logNormCounts(sce)

query_data <- runPCA(query_data)

query_data <- calculateOutlierScoreByCluster(query_data,
                                             clusterVar = "reclustered.broad",
                                             plot = TRUE,
                                             dimred = "PCA",
                                             use_pcs = TRUE)
}
