% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAveragePairwiseCorrelation.R
\name{computeAveragePairwiseCorrelation}
\alias{computeAveragePairwiseCorrelation}
\title{Compute Average Pairwise Correlation between Cell Types}
\usage{
computeAveragePairwiseCorrelation(
  query_data,
  reference_data,
  query_cell_type_col,
  ref_cell_type_col,
  cell_types,
  correlation_method
)
}
\arguments{
\item{query_data}{An object of class "SingleCellExperiment" containing the query cell data.}

\item{reference_data}{An object of class "SingleCellExperiment" containing the reference cell data.}

\item{query_cell_type_col}{A character string specifying the column name for cell types in query_data.}

\item{ref_cell_type_col}{A character string specifying the column name for cell types in ref_data.}

\item{cell_types}{A character vector specifying the cell types to consider.}

\item{correlation_method}{The correlation method to use for calculating pairwise correlations.}
}
\value{
The average pairwise correlation matrix.
}
\description{
This function computes the pairwise correlations between query and reference cells based on the provided correlation method.
It then calculates the average correlation for each pair of cell types.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR
scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- scores$labels

# Compute Pairwise Correlations
## Note: The selection of highly variable genes and desired cell types may vary based on user preference.
## Note: The cell type annotation method used in this example is SingleR. You can use any other method for cell type annotation and provide the corresponding labels in the metadata.
## Here, we are demonstrating the usage with arbitrary choices.

# Selecting highly variable genes
ref_var <- getTopHVGs(ref_data, n = 2000)
query_var <- getTopHVGs(query_data, n = 2000)

# Intersect the gene symbols to obtain common genes
common_genes <- intersect(ref_var, query_var)

# Select desired cell types
selected_cell_types <- c("CD4", "CD8", "B_and_plasma")
ref_data_subset <- ref_data[common_genes, ref_data$reclustered.broad \%in\% selected_cell_types]
query_data_subset <- query_data[common_genes, query_data$reclustered.broad \%in\% selected_cell_types]

# Compute pairwise correlations
cor_matrix_avg <- computeAveragePairwiseCorrelation(query_data = query_data_subset, reference_data = ref_data_subset, query_cell_type_col = "labels", ref_cell_type_col = "reclustered.broad", cell_types = selected_cell_types, correlation_method = "spearman")

# Visualize the results using any visualization method of choice

}
