% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{covid_reference_data}
\alias{covid_reference_data}
\title{COVID-19 PBMC Reference (Healthy) Single-Cell RNA-Seq Dataset}
\format{
An object of class \code{SingleCellExperiment} with 414 rows and 900 columns.
}
\source{
Stephenson E, et al. (2021), processed and hosted at
\url{https://doi.org/10.5281/zenodo.18274942}.
}
\usage{
covid_reference_data
}
\description{
This dataset contains a downsampled reference (healthy
donor) subset of the Stephenson et al. (2021) COVID-19 PBMC atlas. It has
been preprocessed to include log-normalized counts, an author-provided
cell-type column, and PCA results, restricted to 5 shared cell types
(including CD14 monocytes) and a gene panel that always includes the
Yoshida et al. interferon-response signature.
}
\details{
This dataset underwent the following steps:
\itemize{
 \item Downloads the processed reference (healthy) object from Zenodo
 (\url{https://doi.org/10.5281/zenodo.18274942}), itself derived from the
 Stephenson et al. (2021) COVID-19 PBMC atlas (CZI CELLxGENE).
 \item Restricts to 5 shared cell types (CD14 monocytes, CD4 T, CD8 T, B
 cells, NK_16hi) using the \code{author_cell_type_merged} column.
 \item Downsamples to at most 180 cells per cell type.
 \item Selects highly variable genes using the function \code{getTopHVGs}
 from the \code{scran} package, intersected between reference and query,
 always including the 25-gene Yoshida et al. (2019) interferon-response
 signature.
 \item Performs Principal Component Analysis (PCA) using the function
 \code{runPCA} from the \code{scater} package.
}
}
\examples{
# Load and explore the COVID-19 reference dataset
data("covid_reference_data")

}
\references{
Stephenson E, et al. (2021). Single-cell multi-omics analysis
of the immune response in COVID-19. Nature Medicine 27:904-916.
}
\seealso{
Use \code{data("covid_reference_data")} to load and access the
resulting reference dataset, and \code{data("covid_query_data")} for the
corresponding query (severe COVID-19) dataset.
}
\keyword{internal}
