% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regressPCObject.R
\name{.getAvailableCoefficients}
\alias{.getAvailableCoefficients}
\title{Determine Available Coefficient Types}
\usage{
.getAvailableCoefficients(x)
}
\arguments{
\item{x}{An object of class \code{regressPCObject} containing regression
results from \code{regressPC} function.}
}
\value{
A character vector of available coefficient types.
}
\description{
Helper function to determine which coefficient types are
available in a regressPCObject based on the model type and data structure.
}
\author{
Anthony Christidis,
\email{anthony-alexander_christidis@hms.harvard.edu}
}
\keyword{internal}
