% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateColors.R
\name{generateColors}
\alias{generateColors}
\title{Generate Paired Colors for Cell Types}
\usage{
generateColors(cell_type_names, paired = FALSE)
}
\arguments{
\item{cell_type_names}{A character vector of cell type names that need to be
assigned colors.}

\item{paired}{If TRUE, the colored returned should be paired. Default is
FALSE.}
}
\value{
A named character vector where the names are the original cell type
names, and the values are the assigned colors.
}
\description{
This function assigns paired colors (light and dark) to a list
of cell type names. The colors are selected from various color palettes in
the `pals` package.
}
\details{
The function uses color palettes from the `pals` package to generate
colors or pairs of colors (light and dark) for each cell type name provided.
It cycles through different color families (blues, greens, reds, oranges,
purples, purd and greys) to create the colors
}
\author{
Anthony Christidis,
\email{anthony-alexander_christidis@hms.harvard.edu}
}
\keyword{internal}
