% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performLinearRegression.R
\name{performLinearRegression}
\alias{performLinearRegression}
\title{Perform Linear Regression Analysis on Single-Cell Data}
\usage{
performLinearRegression(se_object, dependent_var, independent_var)
}
\arguments{
\item{se_object}{A SingleCellExperiment object containing the data for regression analysis}

\item{dependent_var}{A character string specifying the dependent variable principal component (e.g., "PC1", "PC2", etc.)}

\item{independent_var}{A character string specifying the column name for the independent variable in the SingleCellExperiment object}
}
\value{
The summary of the linear regression model
}
\description{
This function performs linear regression on a SingleCellExperiment object, where the dependent variable is specified by the user as one of the principal components (PC1, PC2, etc.) from the dimension reduction slot, and the independent variable is provided as a column name in the SingleCellExperiment object.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Run PCA
query_data <- runPCA(query_data)

# Get cell type scoresx using SingleR or any other method
scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- pred$labels

# Perform linear regression on PC1 and a specific variable (e.g., "labels")
performLinearRegression(query_data, "PC1", "labels")

# Note: Instead of using SingleR, you can use any other method to obtain the scores for regression analysis.
# Make sure to provide the appropriate dependent and independent variables to the performLinearRegression function.

}
