% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellTypeDistribution.R
\name{plotCellTypeDistribution}
\alias{plotCellTypeDistribution}
\title{Exploring Cell-Type Specific Patterns: Counts per Cell and Annotation Scores Visualization}
\usage{
plotCellTypeDistribution(cell_type_scores, query_data, cell_type)
}
\arguments{
\item{cell_type_scores}{A numeric vector of scores for the specific cell type obtained from a cell type annotation method.}

\item{query_data}{An object of class "SingleCellExperiment" containing a numeric expression matrix.}

\item{cell_type}{A character string specifying the name of the cell type for which the distribution of log-transformed counts per cell is to be visualized.}
}
\value{
A grid.arrange object displaying histograms of log-transformed counts and scores.
        This object can be further customized or used for additional plot manipulations.
}
\description{
Creates histograms of gene expression values and scores obtained from cell type annotation methods for a specific cell type.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)
library(gridExtra)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log-transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR
cell_type_scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)$scores[, "CD4"]

# Visualize the distribution of log-transformed counts and scores
plotCellTypeDistribution(cell_type_scores, query_data, "CD4")

# Note: Users can use any cell type annotation method of their choice to obtain the scores.
# Ensure that the scores and log-transformed counts are provided to the function for visualization.

}
