% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneExpressionDistribution.R
\name{plotGeneExpressionDistribution}
\alias{plotGeneExpressionDistribution}
\title{Plot gene expression distribution from overall and cell type-specific perspective}
\usage{
plotGeneExpressionDistribution(
  query_data,
  cell_type_labels,
  cell_type,
  feature
)
}
\arguments{
\item{cell_type_labels}{A character column name from the colData table of the SingleCellExperiment object representing cell type information.}

\item{cell_type}{A character string representing the name of the cell type for which the specific gene distribution is to be visualized.}

\item{feature}{A character string representing the gene name for which the distribution is to be visualized.}

\item{se_object}{An object of class "SingleCellExperiment" containing log-transformed expression matrix and other metadata.
It can be either a reference or query dataset.}
}
\value{
A ggplot object displaying the histogram plots of gene expression distribution from the overall dataset and cell type-specific perspective.
        This object can be further customized or used for additional plot manipulations.
}
\description{
This function generates histogram plots to visualize the distribution of gene expression values for a specific gene
across the overall dataset and within a specified cell type.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR or any other method
scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- scores$labels

# # Visualize the distribution of log-transformed counts and scores for the se_object
# Note: Users can use SingleR or any other method to obtain the cell type scores.
# Ensure that the scores and log-transformed counts are provided to the function for visualization.
# For demonstration we have used query_data.
plotGeneExpressionDistribution(query_data, "labels", "B_and_plasma", "VPREB3")

}
