% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressPC.R
\name{plotRegressPC}
\alias{plotRegressPC}
\title{Plot PC Regression diagnostic}
\usage{
plotRegressPC(sce, regressPC_res, dep.vars = NULL, indep.var, max_pc = 20)
}
\arguments{
\item{sce}{An object of class \code{\linkS4class{SingleCellExperiment}}
containing the data for regression analysis.}

\item{regressPC_res}{a result from \code{\link{regressPC}}}

\item{dep.vars}{character. Dependent variable(s). Determines which principal
component(s) (e.g., "PC1", "PC2", etc.) are used as explanatory variables.
Principal components are expected to be stored in a PC matrix named
\code{"PCA"} in the \code{reducedDims} of \code{sce}. Defaults to
\code{NULL} which will then regress on each principal component present in
the PC matrix.}

\item{indep.var}{character. Independent variable. A column name in the
\code{colData} of \code{sce} specifying the response variable.}

\item{max_pc}{The maximum number of PCs to show on the plot. Set to 0 to show
all.}
}
\value{
a ggplot2 plot object
}
\description{
Plot the result of a PC regression diagnostic.
}
