% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngquantCompression.R
\name{pngquantCompression}
\alias{pngquantCompression}
\title{Compress and Save a ggplot Object Using pngquant}
\usage{
pngquantCompression(
  plot,
  filename,
  height = 5,
  width = 10,
  quality = "1-5",
  verbose = FALSE
)
}
\arguments{
\item{plot}{A `ggplot` object to be saved and compressed.}

\item{filename}{A character string specifying the file path where the plot will be saved.}

\item{height}{Numeric value indicating the height of the saved plot in inches. Default is 5 inches.}

\item{width}{Numeric value indicating the width of the saved plot in inches. Default is 10 inches.}

\item{quality}{A character string specifying the quality range for compression. The range should be in the form `"min-max"` where `min` and `max` are integers between 0 and 100. Default is `"1-5"`.}

\item{verbose}{Logical; if `TRUE`, additional information about the process is printed. Default is `FALSE`.}
}
\value{
The function does not return a value. It saves and compresses the plot directly to the specified file path.
}
\description{
This function saves a ggplot object to a file and compresses the resulting image using `pngquant`.
}
\details{
The function utilizes `ggsave` to save a `ggplot2` plot to a specified file path. After saving, it runs the `pngquant` command-line tool to compress the image. The compression quality can be controlled via the `quality` parameter. Ensure `pngquant` is installed and properly configured on your system for the compression to work.
}
\author{
Anthony Christidis, \email{anthony-alexander_christidis@hms.harvard.edu}
}
\keyword{internal}
