% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projectDiscriminantSpace.R,
%   R/projectDiscriminantSpace.R
\name{plot.projectDiscriminantSpace}
\alias{plot.projectDiscriminantSpace}
\alias{projectDiscriminantSpace}
\title{Plot Projected Data on Discriminant Spaces}
\usage{
\method{plot}{projectDiscriminantSpace}(x, cell_types, plot_type = c("scatterplot", "boxplot"), ...)

projectDiscriminantSpace(
  reference_data,
  query_data = NULL,
  ref_cell_type_col,
  query_cell_type_col,
  eigen_threshold = 0.1,
  n_tree = 500,
  n_top = 20,
  alpha = 0.01
)
}
\arguments{
\item{x}{An object of class \code{projectDiscriminantSpace} containing the projected data on the discriminant space.. 
Each element of the list represents a combination of cell types and datasets. Each element should contain 'ref_proj' and 'query_proj' data frames.}

\item{cell_types}{A character vector specifying the cell types to plot. If not provided, all cell types will be plotted.}

\item{plot_type}{Type of plot to generate. Options are "scatterplot" and "boxplot". Default is "scatterplot".}

\item{...}{Additional arguments to be passed to the plotting functions.}

\item{reference_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the reference cells.}

\item{query_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the query cells.
If NULL, only the projected reference data is returned. Default is NULL.}

\item{ref_cell_type_col}{The column name in \code{reference_data} indicating cell type labels.}

\item{query_cell_type_col}{The column name in \code{query_data} indicating cell type labels.}

\item{eigen_threshold}{A numeric value specifying the threshold for retaining eigenvalues in discriminant analysis.}

\item{n_tree}{An integer specifying the number of trees for the random forest used in variable importance calculation.}

\item{n_top}{An integer specifying the number of top variables to select based on importance scores.}

\item{alpha}{A numeric value specifying the significance level for Mahalanobis distance cutoff.}
}
\value{
A ggplot object representing the scatterplot or boxplot of the projected data.

A list with the following components for each cell type combination:
\item{discriminant_eigenvalues}{Eigenvalues from the discriminant analysis.}
\item{discriminant_eigenvectors}{Eigenvectors from the discriminant analysis.}
\item{ref_proj}{Reference data projected onto the discriminant space.}
\item{query_proj}{Query data projected onto the discriminant space.}
\item{query_mahalanobis_dist}{Mahalanobis distances of query projections.}
\item{mahalanobis_crit}{Cutoff value for Mahalanobis distance significance.}
\item{query_cosine_similarity}{Cosine similarity scores of query projections.}
}
\description{
This function plots the projected reference and query data on discriminant spaces.

This function projects query single-cell RNA-seq data onto the discriminant space defined by reference data. The reference data 
is used to identify important variables and compute discriminant vectors, which are then used to project both reference and query 
data. Similarity between the query and reference projections is assessed using cosine similarity and Mahalanobis distance.
}
\details{
This function generates either a scatterplot or a boxplot to visualize the projected data onto the discriminant spaces.
For scatterplot, each point represents a projected data point, and colors are used to differentiate between different cell types 
and datasets. For boxplot, the distribution of the projected data values for each cell type is shown, separated by datasets.

The function performs the following steps for each pairwise combination of cell types:
\itemize{
  \item Identifies the top important variables to distinguish the two cell types from the reference data.
  \item Computes the Ledoit-Wolf shrinkage estimate of the covariance matrix for each cell type using the top important genes.
  \item Constructs within-class and between-class scatter matrices.
  \item Solves the generalized eigenvalue problem to obtain discriminant vectors.
  \item Projects both reference and query data onto the discriminant space.
  \item Assesses similarity of the query data projection to the reference data using cosine similarity and Mahalanobis distance.
}
}
\examples{
# Load necessary library
library(scRNAseq)
library(scuttle)
library(SingleR)
library(scran)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR (or any other cell type annotation method)
scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- scores$labels

# Selecting highly variable genes (can be customized by the user)
ref_var <- getTopHVGs(ref_data, n = 500)
query_var <- getTopHVGs(query_data, n = 500)

# Intersect the gene symbols to obtain common genes
common_genes <- intersect(ref_var, query_var)
ref_data_subset <- ref_data[common_genes, ]
query_data_subset <- query_data[common_genes, ]

# Project query data onto discriminant space of reference data 
# for each pairwise combination of cell types
disc_output <- projectDiscriminantSpace(reference_data = ref_data_subset,
                                        query_data = query_data_subset, 
                                        query_cell_type_col = c("reclustered.broad", "labels")[1], 
                                        ref_cell_type_col = "reclustered.broad",
                                        eigen_threshold  = 1e-1,
                                        n_tree = 500,
                                        n_top = 25,
                                        alpha = 0.01)
                                        
# Generate scatter and boxplot
plot(disc_output, plot_type = "scatterplot")
plot(disc_output, cell_types = "CD4-CD8", plot_type = "boxplot")

}
\seealso{
\code{\link{projectDiscriminantSpace}}

\code{\link{plot.projectDiscriminantSpace}}
}
\author{
Anthony Christidis, \email{anthony-alexander_christidis@hms.harvard.edu}
}
