% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHVGOverlap.R
\name{calculateHVGOverlap}
\alias{calculateHVGOverlap}
\title{Calculate the Overlap Coefficient for Highly Variable Genes}
\usage{
calculateHVGOverlap(reference_genes, query_genes)
}
\arguments{
\item{reference_genes}{A vector of highly variable genes from the reference dataset.}

\item{query_genes}{A vector of highly variable genes from the query dataset.}
}
\value{
Overlap coefficient, a value between 0 and 1, where 0 indicates no overlap 
        and 1 indicates complete overlap of highly variable genes between datasets.
}
\description{
This function calculates the overlap coefficient between the sets of highly variable genes 
from a reference dataset and a query dataset. The overlap coefficient measures the 
degree of overlap or similarity between these two sets of genes, pointing towards how well reference
and query datasets are aligned.
}
\examples{

library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Selcting highly variable genes

ref_var <- getTopHVGs(ref_data, n=2000)
query_var <- getTopHVGs(query_data, n=2000)

overlap_coefficient <- calculateHVGOverlap(reference_genes = ref_var, 
                                          query_genes = query_var)
}
