% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePCASubspace.R
\name{comparePCASubspace}
\alias{comparePCASubspace}
\title{Compare Subspaces Spanned by Top Principal Components}
\usage{
comparePCASubspace(
  reference_data,
  query_data,
  pc_subset = c(1:5),
  n_top_vars = 50
)
}
\arguments{
\item{reference_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the reference cells.}

\item{query_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the query cells.}

\item{pc_subset}{A numeric vector specifying the subset of principal components (PCs) to compare. Default is the first five PCs.}

\item{n_top_vars}{An integer indicating the number of top loading variables to consider for each PC. Default is 50.}
}
\value{
A list containing the following components:
  \item{principal_angles_cosines}{A numeric vector of cosine values of principal angles.}
  \item{average_variance_explained}{A numeric vector of average variance explained by each PC.}
  \item{weighted_cosine_similarity}{A numeric value representing the weighted cosine similarity.}
}
\description{
This function compares the subspace spanned by the top principal components (PCs) in a reference dataset to that 
in a query dataset. It computes the cosine similarity between the loadings of the top variables for each PC in 
both datasets and provides a weighted cosine similarity score.
}
\details{
This function compares the subspace spanned by the top principal components (PCs) in a reference dataset 
to that in a query dataset. It first computes the cosine similarity between the loadings of the top variables 
for each PC in both datasets. The top cosine similarity scores are then selected, and their corresponding PC 
indices are stored. Additionally, the function calculates the average percentage of variance explained by the 
selected top PCs. Finally, it computes a weighted cosine similarity score based on the top cosine similarities 
and the average percentage of variance explained.
}
\examples{
# Load necessary library
library(scRNAseq)
library(scuttle)
library(scran)
library(SingleR)
library(ggplot2)
library(scater)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR (or any other cell type annotation method)
scores <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- scores$labels

# Selecting highly variable genes (can be customized by the user)
ref_var <- getTopHVGs(ref_data, n = 500)
query_var <- getTopHVGs(query_data, n = 500)

# Intersect the gene symbols to obtain common genes
common_genes <- intersect(ref_var, query_var)
ref_data_subset <- ref_data[common_genes, ]
query_data_subset <- query_data[common_genes, ]

# Subset reference and query data for a specific cell type
ref_data_subset <- ref_data_subset[, which(ref_data_subset$reclustered.broad == "CD8")]
query_data_subset <- query_data_subset[, which(colData(query_data_subset)$labels == "CD8")]

# Run PCA on the reference and query datasets
ref_data_subset <- runPCA(ref_data_subset, ncomponents = 50)
query_data_subset <- runPCA(query_data_subset, ncomponents = 50)

# Compare PCA subspaces
subspace_comparison <- comparePCASubspace(query_data_subset, ref_data_subset, 
                                          pc_subset = c(1:5), n_top_vars = 50)

# Create a data frame for plotting
plot(subspace_comparison)

}
\seealso{
\code{\link{plot.comparePCASubspace}}
}
\author{
Anthony Christidis, \email{anthony-alexander_christidis@hms.harvard.edu}
}
