% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCategorizationEntropy.R
\name{nElements}
\alias{nElements}
\title{Number of Elements}
\usage{
nElements(X)
}
\arguments{
\item{X}{A matrix or vector.}
}
\value{
An integer representing the number of elements in \code{X}.
}
\description{
This function returns the number of elements in a matrix or vector.
}
\details{
If \code{X} is a matrix, the function returns the product of its dimensions. If \code{X} is a vector, the function returns
its length.
}
\author{
Andrew Ghazi, \email{andrew_ghazi@hms.harvard.edu}
}
\keyword{internal}
