% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetScores.R
\name{plotGeneSetScores}
\alias{plotGeneSetScores}
\title{Visualization of gene sets or pathway scores on dimensional reduction plot}
\usage{
plotGeneSetScores(se_object, method, feature)
}
\arguments{
\item{se_object}{An object of class "SingleCellExperiment" containing numeric expression matrix and other metadata.
It can be either a reference or query dataset.}

\item{method}{A character string indicating the method for visualization ("PCA", "TSNE", or "UMAP").}

\item{feature}{A character string representing the name of the feature (score) in the colData(query_data) to plot.}
}
\value{
A plot object with the scores plotted on the selected dimensional reduction plot.
}
\description{
Plot gene sets or pathway scores on PCA, TSNE, or UMAP. Single cells are color-coded by scores of gene sets or pathways.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(AUCell)

## load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

## log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Run PCA on the query data
query_data <- runPCA(query_data)

# Assign scores to colData (users should ensure that the scores are present in the colData)
colData(query_data)$geneSetScores <- # ... (assign the scores here)

# Plot gene set scores on PCA
plotGeneSetScores(se_object = query_data, method = "PCA", feature = "geneSetScores")

# Note: Users can provide their own gene set scores in the colData of the 'se_object' object, using any method of their choice.
# Ensure that the scores are assigned to the colData and specify the correct feature name for visualization.
The example code shows the assignment of gene set scores to the colData of the query object for demonstration purposes.

}
