% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkerExpression.R
\name{plotMarkerExpression}
\alias{plotMarkerExpression}
\title{Plot gene expression distribution from overall and cell type-specific perspective}
\usage{
plotMarkerExpression(
  reference_data,
  query_data,
  ref_cell_type_col,
  query_cell_type_col,
  gene_name,
  label
)
}
\arguments{
\item{reference_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the reference cells.}

\item{query_data}{A \code{\linkS4class{SingleCellExperiment}} object containing numeric expression matrix for the query cells.}

\item{ref_cell_type_col}{character. The column name in the \code{colData} of \code{reference_data} that identifies the cell types.}

\item{query_cell_type_col}{character. The column name in the \code{colData} of \code{query_data} that identifies the cell types.}

\item{gene_name}{character. A string representing the gene name for which the distribution is to be visualized.}

\item{label}{character. A vector of cell type labels to plot (e.g., c("T-cell", "B-cell")).}
}
\value{
A gtable object containing two arranged density plots as grobs. 
        The first plot shows the overall gene expression distribution, 
        and the second plot displays the cell type-specific expression 
        distribution.
}
\description{
This function generates density plots to visualize the distribution of gene expression values 
for a specific gene across the overall dataset and within a specified cell type.
}
\details{
This function generates density plots to compare the distribution of a specific marker 
gene between reference and query datasets. The aim is to inspect the alignment of gene expression 
levels as a surrogate for dataset similarity. Similar distributions suggest a good alignment, 
while differences may indicate discrepancies or incompatibilities between the datasets.
}
\examples{
\donttest{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR or any other method
pred <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- pred$labels

# Note: Users can use SingleR or any other method to obtain the cell type annotations.
plotMarkerExpression(reference_data = ref_data, 
                     query_data = query_data, 
                     ref_cell_type_col = "reclustered.broad", 
                     query_cell_type_col = "labels", 
                     gene_name = "VPREB3", 
                     label = "B_and_plasma")
}

}
