% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkerExpression.R
\name{plotMarkerExpression}
\alias{plotMarkerExpression}
\title{Plot gene expression distribution from overall and cell type-specific perspective}
\usage{
plotMarkerExpression(
  reference_data,
  query_data,
  reference_cell_labels,
  query_cell_labels,
  gene_name,
  label
)
}
\arguments{
\item{reference_data}{An object of class "SingleCellExperiment" containing log-transformed expression matrix and metadata for the reference dataset.}

\item{query_data}{An object of class "SingleCellExperiment" containing log-transformed expression matrix and metadata for the query dataset.}

\item{reference_cell_labels}{A character column name from the colData(reference_data) of the reference dataset representing cell type information.}

\item{query_cell_labels}{A character column name from the colData(query_data) of the query dataset representing cell type information.}

\item{gene_name}{A character string representing the gene name for which the distribution is to be visualized.}

\item{label}{A character vector of cell type labels to plot (e.g., c("T-cell", "B-cell"))}
}
\value{
A object displaying density plots of gene expression distribution from both the 
        overall dataset and a cell type-specific perspective.
}
\description{
This function generates histogram plots to visualize the distribution of gene expression values for a specific gene
across the overall dataset and within a specified cell type.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR or any other method
pred <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- pred$labels

# Note: Users can use SingleR or any other method to obtain the cell type annotations.
plotMarkerExpression(reference_data = ref_data, 
                     query_data = query_data, 
                     reference_cell_labels = "reclustered.broad", 
                     query_cell_labels = "labels", 
                     gene_name = "VPREB3", 
                     label = "B_and_plasma")

}
