% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCvsAnnotation.R
\name{plotQCvsAnnotation}
\alias{plotQCvsAnnotation}
\title{Scatter plot: QC stats vs Cell Type Annotation Scores}
\usage{
plotQCvsAnnotation(query_data, qc_col, label_col, score_col, label = NULL)
}
\arguments{
\item{query_data}{A \code{\linkS4class{SingleCellExperiment}} containing the single-cell 
expression data and metadata.}

\item{qc_col}{character. A column name in the \code{colData} of \code{query_data} that 
contains the QC stats of interest.}

\item{label_col}{character. The column name in the \code{colData} of \code{query_data} 
that contains the cell type labels.}

\item{score_col}{character. The column name in the \code{colData} of \code{query_data} that 
contains the cell type annotation scores.}

\item{label}{character. A vector of cell type labels to plot (e.g., c("T-cell", "B-cell")).
Defaults to \code{NULL}, which will include all the cells.}
}
\value{
A ggplot object displaying a scatter plot of QC stats vs annotation scores, 
        where each point represents a cell, color-coded by its cell type.
}
\description{
Creates a scatter plot to visualize the relationship between QC stats (e.g., library size) 
and cell type annotation scores for one or more cell types.
}
\details{
This function generates a scatter plot to explore the relationship between various quality 
control (QC) statistics, such as library size and mitochondrial percentage, and cell type 
annotation scores. By examining these relationships, users can assess whether specific QC 
metrics, systematically influence the confidence in cell type annotations, 
which is essential for ensuring reliable cell type annotation.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log-transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR
pred <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)
pred <- as.data.frame(pred)

# Assign labels to query data
colData(query_data)$labels <- pred$labels
scores <- apply(pred[,1:4], 1, max)

# Assign scores to query data
colData(query_data)$cell_scores <- scores

plotQCvsAnnotation(query_data = query_data, 
                  qc_col = "percent.mito", 
                  label_col = "labels", 
                  score_col = "cell_scores", 
                  label = NULL)
                  
# Generate scatter plots
plotQCvsAnnotation(query_data = query_data, 
                   qc_col = "percent.mito", 
                   label_col = "labels", 
                   score_col = "cell_scores", 
                   label = c("CD4", "CD8"))
                   
}
