% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCvsAnnotation.R
\name{plotQCvsAnnotation}
\alias{plotQCvsAnnotation}
\title{Scatter plot: QC stats and Cell Type Scores}
\usage{
plotQCvsAnnotation(query_data, qc_col, label_col, score_col, label = NULL)
}
\arguments{
\item{query_data}{An object of class "SingleCellExperiment" containing the single-cell expression data and metadata.}

\item{qc_col}{A character string specifying the column name in the colData(query_data) that contains the QC stats of interest.}

\item{label_col}{A character string specifying the column name in the colData(query_data) that contains the cell type labels.}

\item{score_col}{A character string specifying the column name in the colData(query_data) that contains the cell type scores.}

\item{label}{A character vector of cell type labels to plot (e.g., c("T-cell", "B-cell")). If NULL, all cells will be included.}
}
\value{
A ggplot object displaying the scatter plot of total UMIs and annotation scores.
}
\description{
Creates a scatter plot to visualize the relationship between QC stats (e.g., library size) and cell type scores for one or more cell types.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# Log-transform datasets
ref_data <- logNormCounts(ref_data)
query_data <- logNormCounts(query_data)

# Get cell type scores using SingleR
pred <- SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)
pred <- as.data.frame(pred)

# Assign labels to query data
colData(query_data)$labels <- pred$labels
scores <- apply(pred[,1:4], 1, max)

# Assign scores to query data
colData(query_data)$cell_scores <- scores

# Generate scatter plots
plotQCvsMultipleCellTypes(query_data, "percent.mito", "labels", "cell_scores", c("CD4", "CD8"))
plotQCvsAnnotation(query_data, "percent.mito", "labels", "cell_scores", c("CD4", "CD8"))
plotQCvsAnnotation(query_data, "percent.mito", "labels", "cell_scores", NULL)

}
