% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPCA.R
\name{projectPCA}
\alias{projectPCA}
\title{Project Query Data Onto PCA Space of Reference Data}
\usage{
projectPCA(
  query_data,
  reference_data,
  n_components = 10,
  query_cell_type_col = NULL,
  ref_cell_type_col = NULL,
  return_value = c("data.frame", "list")[1]
)
}
\arguments{
\item{query_data}{A SingleCellExperiment object containing the data to be projected.}

\item{reference_data}{A SingleCellExperiment object containing the reference data with pre-computed PCA.}

\item{n_components}{An integer specifying the number of principal components to use for projection. Defaults to 10. 
Must be less than or equal to the number of components available in the reference PCA.}

\item{query_cell_type_col}{character. The column name in the \code{colData} of \code{query_data} 
that identifies the cell types.}

\item{ref_cell_type_col}{character. The column name in the \code{colData} of \code{reference_data} 
that identifies the cell types.}

\item{return_value}{A character string specifying the format of the returned data. Can be \code{data.frame} (combined reference 
and query projections) or \code{list} (separate lists for reference and query projections) (default = \code{data.frame}).}
}
\value{
A \code{data.frame} containing the projected data in rows (reference and query data combined) or a \code{list} containing 
separate matrices for reference and query projections, depending on the \code{return_value} argument.
}
\description{
This function projects a query singleCellExperiment object onto the PCA space of a reference 
singleCellExperiment object. The PCA analysis on the reference data is assumed to be pre-computed and stored within the object.
}
\details{
This function assumes that the "PCA" element exists within the \code{reducedDims} of the reference data 
(obtained using \code{reducedDim(reference_data)}) and that the genes used for PCA are present in both the reference and query data. 
It performs centering and scaling of the query data based on the reference data before projection.
}
\examples{
# Load required libraries
library(scRNAseq)
library(scuttle)
library(SingleR)
library(scran)
library(scater)
library(RColorBrewer)

# Load data (replace with your data loading)
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(assay(sce)), size = floor(0.7 * ncol(assay(sce))), replace = FALSE)
ref_data <- sce[, indices]
query_data <- sce[, -indices]

# log transform datasets
ref_data <- scuttle::logNormCounts(ref_data)
query_data <- scuttle::logNormCounts(query_data)

# Get cell type scores using SingleR (or any other cell type annotation method)
scores <- SingleR::SingleR(query_data, ref_data, labels = ref_data$reclustered.broad)

# Add labels to query object
colData(query_data)$labels <- scores$labels

# Selecting highly variable genes (can be customized by the user)
ref_var <- scran::getTopHVGs(ref_data, n = 2000)
query_var <- scran::getTopHVGs(query_data, n = 2000)

# Intersect the gene symbols to obtain common genes
common_genes <- intersect(ref_var, query_var)
ref_data_subset <- ref_data[common_genes, ]
query_data_subset <- query_data[common_genes, ]

# Run PCA on the reference data (assumed to be prepared)
ref_data_subset <- runPCA(ref_data_subset)

# Project the query data onto PCA space of reference
pca_output <- projectPCA(query_data = query_data_subset, reference_data = ref_data_subset,
                         n_components = 10,
                         query_cell_type_col = "labels",
                         ref_cell_type_col = "reclustered.broad",
                         return_value = c("data.frame", "list")[1])

# Compute t-SNE and UMAP using first 10 PCs
tsne_data <- data.frame(calculateTSNE(t(pca_output[, paste0("PC", 1:10)])))
umap_data <- data.frame(calculateUMAP(t(pca_output[, paste0("PC", 1:10)])))

# Combine the cell type labels from both datasets
tsne_data$Type <- paste(pca_output$dataset, pca_output$cell_type)

# Define the cell types and legend order
legend_order <- c("Query CD8",
                  "Reference CD8",
                  "Query CD4",
                  "Reference CD4",
                  "Query B_and_plasma",
                  "Reference B_and_plasma")

# Define the colors for cell types
color_palette <- brewer.pal(length(legend_order), "Paired")
color_mapping <- setNames(color_palette, legend_order)
cell_type_colors <- color_mapping[legend_order]

# Visualize t-SNE output
tsne_plot <- ggplot(tsne_data[tsne_data$Type \%in\% legend_order,],
                    aes(x = TSNE1, y = TSNE2, color = factor(Type, levels = legend_order))) +
    geom_point(alpha = 0.5, size = 1) +
    scale_color_manual(values = cell_type_colors) +
    theme_bw() +
    guides(color = guide_legend(title = "Cell Types"))

}
\author{
Anthony Christidis, \email{anthony-alexander_christidis@hms.harvard.edu}
}
