% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressPC.R
\name{regressPC}
\alias{regressPC}
\title{Principal component regression}
\usage{
regressPC(sce, dep.vars = NULL, indep.var, ...)
}
\arguments{
\item{sce}{An object of class
\code{\linkS4class{SingleCellExperiment}} containing the data
for regression analysis.}

\item{dep.vars}{character. Dependent variable(s). Determines which
principal component(s) (e.g., "PC1", "PC2", etc.) are used as
explanatory variables.  Principal components are expected to be
stored in a PC matrix named \code{"PCA"} in the
\code{reducedDims} of \code{sce}. Defaults to \code{NULL} which
will then regress on each principal component present in the PC
matrix.}

\item{indep.var}{character. Independent variable. A column name in
the \code{colData} of \code{sce} specifying the response
variable.}
}
\value{
A \code{list} containing \itemize{ \item summaries of the
    linear regression models for each specified principal
    component, \item the corresponding R-squared (R2) values, \item
    the variance contributions for each principal component, and
    \item the total variance explained.}
}
\description{
This function performs linear regression of a covariate of interest
onto one or more principal components, based on the data in a
SingleCellExperiment object.
}
\details{
Principal component regression, derived from PCA, can be
    used to quantify the variance explained by a covariate
    interest. Applications for single-cell analysis include
    quantification of batch removal, assessing clustering
    homogeneity, and evaluation of alignment of query and reference
    datasets in cell type annotation settings.  Briefly, the R^2 is
    calculated from a linear regression of the covariate B of
    interest onto each principal component. The variance
    contribution of the covariate effect per principal component is
    then calculated as the product of the variance explained by the
    ith principal component (PC) and the corresponding R2(PCi|B).
    The sum across all variance contributions by the covariate
    effects in all principal components gives the total variance
    explained by the covariate as follows:

    Var(C|B) = sum_{i=1}^G Var(C|PC_i) * R^2 (PC_i | B)

    where, Var(C|PCi) is the variance of the data matrix C explained
    by the ith principal component. See references.
}
\examples{
library(scater)
library(scran)
library(scRNAseq)
library(SingleR)

# Load data
sce <- HeOrganAtlasData(tissue = c("Marrow"), ensembl = FALSE)

# Divide the data into reference and query datasets
set.seed(100)
indices <- sample(ncol(sce),
    size = floor(0.7 * ncol(sce)),
    replace = FALSE
)
ref <- sce[, indices]
query <- sce[, -indices]

# log transform datasets
ref <- logNormCounts(ref)
query <- logNormCounts(query)

# Run PCA
query <- runPCA(query)

# Get cell type scores using SingleR
# Note: replace when using cell type annotation scores from other methods
scores <- SingleR(query, ref, labels = ref$reclustered.broad)

# Add labels to query object
query$labels <- scores$labels

# Specify the dependent variables (principal components) and
# independent variable (e.g., "labels")
dep.vars <- paste0("PC", 1:3)
indep.var <- "labels"

# Perform linear regression on multiple principal components
res <- regressPC(
    sce = query,
    dep.vars = dep.vars,
    indep.var = indep.var
)

# Obtain linear regression summaries and R-squared values
res$regression.summaries
res$rsquared

}
\references{
Luecken et al. Benchmarking atlas-level data
    integration in single-cell genomics. Nature Methods, 19:41-50,
    2022.
}
